preskool-erp/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
psycopg[binary]==3.2.9
aiosqlite==0.21.0
alembic==1.16.4
pydantic==2.11.7
pydantic-settings==2.10.1
eval-type-backport==0.2.2
email-validator==2.3.0
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
orjson==3.11.3
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==1.1.0
